+++ src/main/java/view/Intent.java
@@ -1,4 +1,4 @@
-package controller;
+package view;
 
 import java.util.Arrays;
